* Fixed ambiguous call to pow.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@5679 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ player.cpp
@@ -4187,7 +4187,7 @@ double Player::getLostPercent() const
 		lossPercent *= 0.7;
 
 	std::bitset<5> bitset(blessings);
-	return lossPercent * pow(0.92, bitset.count()) / 100;
+	return lossPercent * pow(0.92, (int)bitset.count()) / 100;
 }
 
 void Player::learnInstantSpell(const std::string& name)
